refactor(cli): fold root.rs into mod.rs
After the tuple-form symmetry fix, `cli/root.rs` holds only:

- The `Cli` struct (entry type)
- The `Command` dispatch enum (thin: every variant is one line)
- 3 cross-cutting help/banner tests

All of which belong naturally at the module's public face. The file
adds a layer of indirection (`mod.rs` re-exports `Cli` and `Command`
from `root.rs`) that doesn't earn its keep.

This commit folds everything in `root.rs` into `mod.rs` and deletes
the now-empty file. `mod.rs` becomes:

- `HELP_STYLES` / `BANNER` constants (unchanged)
- `pub mod agent; pub mod cleanup; pub mod config; pub mod workspace;`
- `pub use` of `AuthCommand`, `ConfigCommand`, `MountCommand`,
  `TrustCommand`, `WorkspaceCommand` (same as before; now re-exported
  from sibling files instead of from the now-deleted `root`)
- `pub struct Cli` + `pub enum Command` (moved from `root.rs` verbatim)
- `mod tests` (moved verbatim)

Import adjustments:

- `src/cli/workspace.rs::tests` — `use super::super::root::{Cli,
  Command};` → `use super::super::{Cli, Command};` (the types now
  live one level up directly in `cli`, not under a `root` sub-module).
- Same for `src/cli/config.rs::tests`.
- `src/cli/mod.rs` — removed the `pub use root::{Cli, Command};`
  re-export; both are now defined here, no re-export needed. The
  `pub mod root;` declaration is also dropped.

Zero behavior change, zero signature change, zero test change.

### Why this mirrors the other module layouts

- `src/workspace/mod.rs` holds the public types (`MountConfig`,
  `WorkspaceConfig`, `WorkspaceEdit`) + submodule declarations +
  re-exports, and the domain logic lives in sibling files.
- `src/config/mod.rs` holds `AppConfig` + the submodule declarations.
- `src/runtime/mod.rs` is tiny — re-exports only.

`src/cli/mod.rs` now matches the `workspace/` pattern: the module
root owns the public entry types (`Cli`, `Command`) + declarations +
re-exports; the per-command-family files hold the details.

### Verification

- `cargo fmt -- --check` — pass
- `cargo clippy` — zero warnings
- `cargo nextest run` — 427 passed, 0 skipped
- `cargo run --bin jackin -- --help` + all subcommand `--help`
  spot-checks — identical output

Net LOC: mod.rs +150, root.rs -151, workspace.rs/config.rs -2 each
(import adjustment). One file removed from the directory.
