+++ src/cli/config.rs
@@ -153,7 +153,7 @@ Examples:
 
 #[cfg(test)]
 mod tests {
-    use super::super::root::{Cli, Command};
+    use super::super::{Cli, Command};
     use super::*;
     use clap::Parser;
